requirements.txt
python-telegram-bot==20.0
pytube==12.1.2